Fix type errors, remove unused code from curation bar (formerly 'footer')
